+++ include/engine/utils/MathIncludes.h
@@ -1,0 +1,10 @@
+#ifndef MATHINCLUDES_H
+#define MATHINCLUDES_H
+
+#include "Matrice.h"
+#include "Point.h"
+#include "List.h"
+#include "Vertex.h"
+#include "Color.h"
+
+#endif // !MATHINCLUDES_H
